+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.43.0] - 2018-07-19
 ### Added
 - HashService with bcrypt implementation
 - SymmetricEncryptionService with AES 256 GCM implementation